requirements.txt
cohere==5.3.3
deepeval
hydra-core 
llama-index-core
llama-index-embeddings-cohere
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-embeddings-huggingface
ipywidgets
chromadb
pandas<2.0.0
pymupdf
jsonlines
seaborn
bitsandbytes
accelerate
spacy
scikit-learn
jsonlines
filelock
nltk
dataclasses
sacrebleu==2.4.0
rouge_score==0.1.2
# flash-attn
einops
sentencepiece
deepspeed
peft
evaluate
spacy
requests==2.32.3
tqdm==4.66.4
tree_sitter==0.21.3
ray==2.32.0
shortuuid==1.0.13
mistralai==0.4.2
anthropic==0.31.1
tree-sitter-java==0.21.0
tree-sitter-javascript==0.21.4